requirements.txt
dj-database-url==0.5.0
Django==1.11.2
djangorestframework==3.7.7
gunicorn==19.7.1
mysqlclient==1.3.12
psycopg2==2.7.4
pytz==2018.3
whitenoise==3.3.1
